requirements.txt
amqp==2.5.2
billiard==3.6.1.0
bleach==2.1.4
celery==4.4.0
certifi==2019.9.11
chardet==3.0.4
Django==2.2.7
django-appconf==1.0.3
django-appcore==0.0.9
django-crispy-contact-form==0.3.8
django-crispy-forms==1.8.1
django-guardian==2.1.0
django-ipware==2.1.0
django-publications==1.0.0
django-recaptcha==2.0.5
html5lib==1.0.1
idna==2.8
importlib-metadata==1.4.0
kombu==4.6.7
Markdown==3.1.1
martor==1.4.4
more-itertools==8.1.0
Pillow==6.2.1
psycopg2==2.8.4
pytz==2019.3
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
vine==1.3.0
webencodings==0.5.1
zipp==0.6.0
django-tempus-dominus==5.1.2.9
